Split oversized files: handlers.go, bootstrap.go, config.go
Split three large files in gestaltd into smaller, focused files to improve
navigability. handlers.go (1916 lines) is split into handlers.go (547),
handlers_auth.go (378), handlers_oauth.go (305), handlers_connect.go (408),
and handlers_tokens.go (335). bootstrap.go (1596 lines) is split into
bootstrap.go (503), provider_build.go (937), and secret_resolution.go (184).
config.go (1389 lines) is split into config.go (877) and config_validate.go
(524).

This is a pure file reorganization with no logic changes. Each function appears
in exactly one file, and file-level types, constants, and variables that are
only used by moved functions travel with them. Shared types and helpers remain
in the original files.
