+++ gestaltd/internal/bootstrap/secret_resolution.go
@@ -1,0 +1,184 @@
+package bootstrap
+
+import (
+	"context"
+	"fmt"
+	"reflect"
+	"strings"
+
+	"github.com/valon-technologies/gestalt/server/core"
+	"github.com/valon-technologies/gestalt/server/internal/config"
+	"gopkg.in/yaml.v3"
+)
+
+const secretPrefix = "secret://"
+
+// resolveSecretRefs walks the config struct and replaces any string value
+// starting with "secret://" with the resolved secret value. The SecretsConfig
+// node is skipped to avoid self-referential resolution.
+func resolveSecretRefs(ctx context.Context, cfg *config.Config, sm core.SecretManager) error {
+	resolve := func(val string) (string, error) {
+		name, ok := strings.CutPrefix(val, secretPrefix)
+		if !ok {
+			return val, nil
+		}
+		resolved, err := sm.GetSecret(ctx, name)
+		if err != nil {
+			return "", &core.SecretResolutionError{Name: name, Err: err}
+		}
+		if resolved == "" {
+			return "", &core.SecretResolutionError{
+				Name: name,
+				Err:  fmt.Errorf("resolved to empty value"),
+			}
+		}
+		return resolved, nil
+	}
+
+	if err := resolveStringFields(&cfg.Server, resolve); err != nil {
+		return err
+	}
+	for name := range cfg.Integrations {
+		intg := cfg.Integrations[name]
+		if err := resolveStringFields(&intg, resolve); err != nil {
+			return err
+		}
+		if intg.Plugin != nil {
+			if err := resolveStringFields(intg.Plugin, resolve); err != nil {
+				return err
+			}
+			for _, conn := range intg.Plugin.Connections {
+				if conn != nil {
+					if err := resolveStringFields(conn, resolve); err != nil {
+						return err
+					}
+				}
+			}
+			if err := resolveYAMLNode(&intg.Plugin.Config, resolve); err != nil {
+				return err
+			}
+		}
+		cfg.Integrations[name] = intg
+	}
+	if cfg.Auth.Provider != nil {
+		if err := resolveStringFields(cfg.Auth.Provider, resolve); err != nil {
+			return err
+		}
+	}
+	if cfg.Datastore.Provider != nil {
+		if err := resolveStringFields(cfg.Datastore.Provider, resolve); err != nil {
+			return err
+		}
+	}
+	if cfg.UI.Plugin != nil {
+		if err := resolveStringFields(cfg.UI.Plugin, resolve); err != nil {
+			return err
+		}
+	}
+
+	// Skip cfg.Secrets.Config to avoid self-referential resolution.
+	if err := resolveYAMLNode(&cfg.Auth.Config, resolve); err != nil {
+		return err
+	}
+	if err := resolveYAMLNode(&cfg.Datastore.Config, resolve); err != nil {
+		return err
+	}
+	if err := resolveYAMLNode(&cfg.Telemetry.Config, resolve); err != nil {
+		return err
+	}
+	if err := resolveYAMLNode(&cfg.Audit.Config, resolve); err != nil {
+		return err
+	}
+
+	return nil
+}
+
+func resolveStringFields(ptr any, resolve func(string) (string, error)) error {
+	v := reflect.ValueOf(ptr).Elem()
+	t := v.Type()
+	for i := 0; i < t.NumField(); i++ {
+		field := v.Field(i)
+		switch field.Kind() {
+		case reflect.String:
+			if !field.CanSet() {
+				continue
+			}
+			resolved, err := resolve(field.String())
+			if err != nil {
+				return err
+			}
+			field.SetString(resolved)
+		case reflect.Struct:
+			if field.CanSet() {
+				if err := resolveStringFields(field.Addr().Interface(), resolve); err != nil {
+					return err
+				}
+			}
+		case reflect.Pointer:
+			if !field.IsNil() && field.Elem().Kind() == reflect.Struct {
+				if err := resolveStringFields(field.Interface(), resolve); err != nil {
+					return err
+				}
+			}
+		case reflect.Map:
+			if field.Type().Key().Kind() == reflect.String && field.Type().Elem().Kind() == reflect.String {
+				for _, k := range field.MapKeys() {
+					resolved, err := resolve(field.MapIndex(k).String())
+					if err != nil {
+						return err
+					}
+					field.SetMapIndex(k, reflect.ValueOf(resolved))
+				}
+			}
+		case reflect.Slice:
+			switch field.Type().Elem().Kind() {
+			case reflect.String:
+				for j := 0; j < field.Len(); j++ {
+					elem := field.Index(j)
+					resolved, err := resolve(elem.String())
+					if err != nil {
+						return err
+					}
+					elem.SetString(resolved)
+				}
+			case reflect.Struct:
+				for j := 0; j < field.Len(); j++ {
+					if err := resolveStringFields(field.Index(j).Addr().Interface(), resolve); err != nil {
+						return err
+					}
+				}
+			}
+		}
+	}
+	return nil
+}
+
+func resolveYAMLNode(node *yaml.Node, resolve func(string) (string, error)) error {
+	if node == nil {
+		return nil
+	}
+	switch node.Kind {
+	case yaml.ScalarNode:
+		if node.Tag == "!!str" || node.Tag == "" {
+			resolved, err := resolve(node.Value)
+			if err != nil {
+				return err
+			}
+			node.Value = resolved
+		}
+	case yaml.MappingNode:
+		// Content is [key, value, key, value, ...]; only resolve values.
+		for i := 1; i < len(node.Content); i += 2 {
+			if err := resolveYAMLNode(node.Content[i], resolve); err != nil {
+				return err
+			}
+		}
+	case yaml.SequenceNode, yaml.DocumentNode:
+		for _, child := range node.Content {
+			if err := resolveYAMLNode(child, resolve); err != nil {
+				return err
+			}
+		}
+	}
+	return nil
+}
