+++ gestaltd/internal/server/handlers.go
@@ -7,24 +7,16 @@ import (
 	"fmt"
 	"log/slog"
 	"net/http"
-	"net/url"
 	"regexp"
 	"sort"
-	"strings"
 	"time"
 
 	"github.com/go-chi/chi/v5"
-	"github.com/google/uuid"
 	"github.com/valon-technologies/gestalt/server/core"
 	"github.com/valon-technologies/gestalt/server/internal/apiexec"
 	"github.com/valon-technologies/gestalt/server/internal/bootstrap"
 	"github.com/valon-technologies/gestalt/server/internal/config"
-	"github.com/valon-technologies/gestalt/server/internal/discovery"
 	"github.com/valon-technologies/gestalt/server/internal/invocation"
-	"github.com/valon-technologies/gestalt/server/internal/metricutil"
-	"github.com/valon-technologies/gestalt/server/internal/oauth"
-	"github.com/valon-technologies/gestalt/server/internal/paraminterp"
-	"github.com/valon-technologies/gestalt/server/internal/principal"
 
 	"google.golang.org/grpc/codes"
 	grpcstatus "google.golang.org/grpc/status"
@@ -43,36 +35,10 @@ var (
 	errResolveUser      = errors.New("failed to resolve user")
 )
 
-func (s *Server) resolveUserID(w http.ResponseWriter, r *http.Request) (string, error) {
-	user := UserFromContext(r.Context())
-	if user == nil || user.Email == "" {
-		writeError(w, http.StatusUnauthorized, "not authenticated")
-		return "", errNotAuthenticated
-	}
-	if id := UserIDFromContext(r.Context()); id != "" {
-		return id, nil
-	}
-	dbUser, err := s.datastore.FindOrCreateUser(r.Context(), user.Email)
-	if err != nil || dbUser == nil || dbUser.ID == "" {
-		writeError(w, http.StatusInternalServerError, "failed to resolve user")
-		return "", errResolveUser
-	}
-	return dbUser.ID, nil
-}
-
-func (s *Server) healthCheck(w http.ResponseWriter, _ *http.Request) {
-	writeJSON(w, http.StatusOK, map[string]string{"status": "ok"})
-}
-
-func (s *Server) readinessCheck(w http.ResponseWriter, _ *http.Request) {
-	if s.readiness != nil {
-		if reason := s.readiness(); reason != "" {
-			writeJSON(w, http.StatusServiceUnavailable, map[string]string{"status": reason})
-			return
-		}
-	}
-	writeJSON(w, http.StatusOK, map[string]string{"status": "ok"})
-}
+var (
+	safeParamValue         = regexp.MustCompile(`^[a-zA-Z0-9._-]+$`)
+	safeTokenResponseValue = regexp.MustCompile(`^[a-zA-Z0-9._:/-]+$`)
+)
 
 type instanceInfo struct {
 	Name       string `json:"name"`
@@ -109,6 +75,37 @@ type connectionParamInfo struct {
 	Required    bool   `json:"required,omitempty"`
 	Description string `json:"description,omitempty"`
 	Default     string `json:"default,omitempty"`
+}
+
+func (s *Server) resolveUserID(w http.ResponseWriter, r *http.Request) (string, error) {
+	user := UserFromContext(r.Context())
+	if user == nil || user.Email == "" {
+		writeError(w, http.StatusUnauthorized, "not authenticated")
+		return "", errNotAuthenticated
+	}
+	if id := UserIDFromContext(r.Context()); id != "" {
+		return id, nil
+	}
+	dbUser, err := s.datastore.FindOrCreateUser(r.Context(), user.Email)
+	if err != nil || dbUser == nil || dbUser.ID == "" {
+		writeError(w, http.StatusInternalServerError, "failed to resolve user")
+		return "", errResolveUser
+	}
+	return dbUser.ID, nil
+}
+
+func (s *Server) healthCheck(w http.ResponseWriter, _ *http.Request) {
+	writeJSON(w, http.StatusOK, map[string]string{"status": "ok"})
+}
+
+func (s *Server) readinessCheck(w http.ResponseWriter, _ *http.Request) {
+	if s.readiness != nil {
+		if reason := s.readiness(); reason != "" {
+			writeJSON(w, http.StatusServiceUnavailable, map[string]string{"status": reason})
+			return
+		}
+	}
+	writeJSON(w, http.StatusOK, map[string]string{"status": "ok"})
 }
 
 func (s *Server) listIntegrations(w http.ResponseWriter, r *http.Request) {
@@ -547,1370 +544,4 @@ func safeOperationErrorMessage(err error) (string, bool) {
 	default:
 		return "", false
 	}
-}
-
-type loginRequest struct {
-	State        string `json:"state"`
-	CallbackPort int    `json:"callback_port,omitempty"`
-}
-
-type authInfoResponse struct {
-	Provider       string `json:"provider"`
-	DisplayName    string `json:"display_name"`
-	LoginSupported bool   `json:"login_supported"`
-}
-
-func (s *Server) authProviderName() string {
-	if s.auth == nil {
-		return "none"
-	}
-	return s.auth.Name()
-}
-
-func (s *Server) authEnabled() bool {
-	return s.auth != nil && !s.noAuth
-}
-
-func (s *Server) authInfo(w http.ResponseWriter, _ *http.Request) {
-	provider := s.authProviderName()
-	displayName := provider
-	if s.auth != nil {
-		if dn, ok := s.auth.(AuthProviderDisplayName); ok {
-			displayName = dn.DisplayName()
-		}
-	}
-	writeJSON(w, http.StatusOK, authInfoResponse{
-		Provider:       provider,
-		DisplayName:    displayName,
-		LoginSupported: s.authEnabled(),
-	})
-}
-
-func (s *Server) startLogin(w http.ResponseWriter, r *http.Request) {
-	startedAt := time.Now()
-	auditAllowed := false
-	auditErr := errors.New("login start failed")
-	providerName := s.authProviderName()
-	defer func() {
-		metricutil.RecordAuthMetrics(r.Context(), startedAt, providerName, "begin_login", auditErr != nil)
-		s.auditHTTPEvent(r.Context(), nil, s.authProviderName(), "auth.login.start", auditAllowed, auditErr)
-	}()
-
-	var req loginRequest
-	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
-		auditErr = errors.New("invalid JSON body")
-		writeError(w, http.StatusBadRequest, "invalid JSON body")
-		return
-	}
-	state := req.State
-	if req.CallbackPort > 0 && req.CallbackPort <= maxPort {
-		state = fmt.Sprintf("%s%d:%s", cliStatePrefix, req.CallbackPort, req.State)
-	}
-	if !s.authEnabled() {
-		auditErr = errors.New("auth is disabled")
-		writeError(w, http.StatusNotFound, "auth is disabled")
-		return
-	}
-	url, err := loginURLForRequest(r.Context(), s.auth, state)
-	if err != nil {
-		auditErr = errors.New("failed to generate login URL")
-		writeError(w, http.StatusInternalServerError, "failed to generate login URL")
-		return
-	}
-	loginURL, err := s.resolvePublicURL(r, url)
-	if err != nil {
-		auditErr = errors.New("failed to resolve login URL")
-		writeError(w, http.StatusInternalServerError, "failed to resolve login URL")
-		return
-	}
-	if s.encryptor != nil {
-		encoded, encErr := encodeLoginState(s.encryptor, loginState{
-			State:     req.State,
-			ExpiresAt: s.now().Add(loginStateTTL).Unix(),
-		})
-		if encErr != nil {
-			auditErr = errors.New("failed to encode login state")
-			writeError(w, http.StatusInternalServerError, "failed to encode login state")
-			return
-		}
-		http.SetCookie(w, &http.Cookie{
-			Name:     loginStateCookieName,
-			Value:    encoded,
-			Path:     "/",
-			MaxAge:   int(loginStateTTL.Seconds()),
-			HttpOnly: true,
-			Secure:   s.secureCookies,
-			SameSite: http.SameSiteLaxMode,
-		})
-	}
-	auditAllowed = true
-	auditErr = nil
-	writeJSON(w, http.StatusOK, map[string]string{"url": loginURL})
-}
-
-func loginURLForRequest(ctx context.Context, provider core.AuthProvider, state string) (string, error) {
-	if providerWithContext, ok := provider.(core.LoginURLContextProvider); ok {
-		return providerWithContext.LoginURLContext(ctx, state)
-	}
-	return provider.LoginURL(state)
-}
-
-func (s *Server) resolvePublicURL(r *http.Request, raw string) (string, error) {
-	if raw == "" {
-		return "", nil
-	}
-	parsed, err := url.Parse(raw)
-	if err != nil {
-		return "", err
-	}
-	if parsed.IsAbs() {
-		return parsed.String(), nil
-	}
-
-	base := s.publicBaseURL
-	if base == "" {
-		scheme := "http"
-		if r.TLS != nil {
-			scheme = "https"
-		}
-		base = scheme + "://" + r.Host
-	}
-	baseURL, err := url.Parse(base)
-	if err != nil {
-		return "", err
-	}
-	if baseURL.Path == "" {
-		baseURL.Path = "/"
-	}
-	return baseURL.ResolveReference(parsed).String(), nil
-}
-
-// AuthProviderDisplayName is an optional interface for a human-readable login label.
-type AuthProviderDisplayName interface {
-	DisplayName() string
-}
-
-// SessionTokenIssuer is an optional interface that auth providers can implement
-// to issue session tokens after login.
-type SessionTokenIssuer interface {
-	IssueSessionToken(identity *core.UserIdentity) (string, error)
-}
-
-// SessionTokenTTLProvider is an optional interface that auth providers can
-// implement to expose their configured session TTL for cookie MaxAge.
-type SessionTokenTTLProvider interface {
-	SessionTokenTTL() time.Duration
-}
-
-func (s *Server) setSessionCookie(w http.ResponseWriter, token string) {
-	maxAge := int(defaultSessionCookieTTL.Seconds())
-	if p, ok := s.auth.(SessionTokenTTLProvider); ok {
-		maxAge = int(p.SessionTokenTTL().Seconds())
-	}
-	http.SetCookie(w, &http.Cookie{
-		Name:     sessionCookieName,
-		Value:    token,
-		Path:     "/",
-		MaxAge:   maxAge,
-		HttpOnly: true,
-		Secure:   s.secureCookies,
-		SameSite: http.SameSiteLaxMode,
-	})
-}
-
-func (s *Server) clearSessionCookie(w http.ResponseWriter) {
-	http.SetCookie(w, &http.Cookie{
-		Name:     sessionCookieName,
-		Value:    "",
-		Path:     "/",
-		MaxAge:   -1,
-		HttpOnly: true,
-		Secure:   s.secureCookies,
-		SameSite: http.SameSiteLaxMode,
-	})
-}
-
-// StatefulCallbackHandler is an optional interface for auth providers that need
-// the OAuth state parameter during callback (e.g., for PKCE where the
-// code_verifier is encrypted in the state).
-type StatefulCallbackHandler interface {
-	HandleCallbackWithState(ctx context.Context, code, state string) (*core.UserIdentity, string, error)
-}
-
-// RequestCallbackHandler is an optional interface for auth providers that need
-// the full callback query map. This is used by executable auth plugins so the
-// host can preserve callback state and provider-specific query parameters.
-type RequestCallbackHandler interface {
-	HandleCallbackRequest(ctx context.Context, query url.Values) (*core.UserIdentity, string, error)
-}
-
-func (s *Server) loginCallback(w http.ResponseWriter, r *http.Request) {
-	startedAt := time.Now()
-	auditAllowed := false
-	auditErr := errors.New("login callback failed")
-	auditUserID := ""
-	providerName := s.authProviderName()
-	defer func() {
-		metricutil.RecordAuthMetrics(r.Context(), startedAt, providerName, "complete_login", auditErr != nil)
-		if auditUserID != "" {
-			s.auditHTTPEventWithUserID(r.Context(), auditUserID, principal.SourceSession.String(), s.authProviderName(), "auth.login.complete", auditAllowed, auditErr)
-			return
-		}
-		s.auditHTTPEvent(r.Context(), nil, s.authProviderName(), "auth.login.complete", auditAllowed, auditErr)
-	}()
-
-	code := r.URL.Query().Get("code")
-	if code == "" {
-		auditErr = errors.New("missing code parameter")
-		writeError(w, http.StatusBadRequest, "missing code parameter")
-		return
-	}
-	if !s.authEnabled() {
-		auditErr = errors.New("auth is disabled")
-		writeError(w, http.StatusNotFound, "auth is disabled")
-		return
-	}
-
-	var identity *core.UserIdentity
-	var originalState string
-	var err error
-
-	if handler, ok := s.auth.(RequestCallbackHandler); ok {
-		identity, originalState, err = handler.HandleCallbackRequest(r.Context(), r.URL.Query())
-	} else if stateful, ok := s.auth.(StatefulCallbackHandler); ok {
-		state := r.URL.Query().Get("state")
-		identity, originalState, err = stateful.HandleCallbackWithState(r.Context(), code, state)
-	} else {
-		originalState = r.URL.Query().Get("state")
-		identity, err = s.auth.HandleCallback(r.Context(), code)
-	}
-	if err != nil {
-		auditErr = errors.New("login failed")
-		slog.ErrorContext(r.Context(), "login callback failed", "error", err)
-		writeError(w, http.StatusUnauthorized, "login failed")
-		return
-	}
-
-	if csrfErr := s.validateLoginState(r, originalState); csrfErr != nil {
-		auditErr = errors.New("login state validation failed")
-		slog.ErrorContext(r.Context(), "login state validation failed", "error", csrfErr)
-		writeError(w, http.StatusForbidden, "login state validation failed")
-		return
-	}
-	if s.encryptor != nil {
-		s.clearLoginStateCookie(w)
-	}
-
-	if r.URL.Query().Get("cli") == "1" {
-		dbUser, dbErr := s.datastore.FindOrCreateUser(r.Context(), identity.Email)
-		if dbErr != nil || dbUser == nil || dbUser.ID == "" {
-			auditErr = errors.New("failed to resolve user")
-			writeError(w, http.StatusInternalServerError, "failed to resolve user")
-			return
-		}
-		auditUserID = dbUser.ID
-		apiToken, plaintext, issueErr := s.issueAPIToken(r.Context(), dbUser.ID, cliLoginTokenName, "", true)
-		if issueErr != nil {
-			auditErr = errors.New("failed to issue CLI API token")
-			writeError(w, http.StatusInternalServerError, "failed to issue CLI API token")
-			return
-		}
-		auditAllowed = true
-		auditErr = nil
-		writeJSON(w, http.StatusOK, createTokenResponse{
-			ID:        apiToken.ID,
-			Name:      apiToken.Name,
-			Token:     plaintext,
-			ExpiresAt: apiToken.ExpiresAt,
-		})
-		return
-	}
-
-	if identity != nil && identity.Email != "" {
-		dbUser, auditPrincipalErr := s.datastore.FindOrCreateUser(r.Context(), identity.Email)
-		switch {
-		case auditPrincipalErr != nil:
-			slog.WarnContext(r.Context(), "login audit user resolution failed", "error", auditPrincipalErr)
-		case dbUser != nil && dbUser.ID != "":
-			auditUserID = dbUser.ID
-		default:
-			slog.WarnContext(r.Context(), "login audit user resolution failed", "error", "authenticated principal missing user ID")
-		}
-	}
-
-	resp := map[string]any{
-		"email":        identity.Email,
-		"display_name": identity.DisplayName,
-	}
-
-	token, err := s.issueSessionToken(identity)
-	if err != nil {
-		auditErr = errors.New("failed to issue session token")
-		writeError(w, http.StatusInternalServerError, "failed to issue session token")
-		return
-	}
-	s.setSessionCookie(w, token)
-
-	auditAllowed = true
-	auditErr = nil
-	writeJSON(w, http.StatusOK, resp)
-}
-
-func (s *Server) validateLoginState(r *http.Request, originalState string) error {
-	if s.encryptor == nil {
-		return nil
-	}
-	cookie, err := r.Cookie(loginStateCookieName)
-	if err != nil {
-		return fmt.Errorf("missing login state cookie")
-	}
-	expected, err := decodeLoginState(s.encryptor, cookie.Value, s.now())
-	if err != nil {
-		return fmt.Errorf("invalid login state cookie: %w", err)
-	}
-	if expected.State != originalState {
-		return fmt.Errorf("login state mismatch")
-	}
-	return nil
-}
-
-func (s *Server) clearLoginStateCookie(w http.ResponseWriter) {
-	http.SetCookie(w, &http.Cookie{
-		Name:     loginStateCookieName,
-		Value:    "",
-		Path:     "/",
-		MaxAge:   -1,
-		HttpOnly: true,
-		Secure:   s.secureCookies,
-		SameSite: http.SameSiteLaxMode,
-	})
-}
-
-type startOAuthRequest struct {
-	Integration      string            `json:"integration"`
-	Connection       string            `json:"connection"`
-	Instance         string            `json:"instance"`
-	Scopes           []string          `json:"scopes"`
-	ConnectionParams map[string]string `json:"connection_params"`
-}
-
-func (s *Server) startIntegrationOAuth(w http.ResponseWriter, r *http.Request) {
-	startedAt := time.Now()
-	auditAllowed := false
-	auditErr := errors.New("oauth start failed")
-	providerName := ""
-	metricProviderName := metricutil.UnknownAttrValue
-	connectionMode := metricutil.UnknownAttrValue
-	defer func() {
-		metricutil.RecordConnectionAuthMetrics(r.Context(), startedAt, metricProviderName, "oauth", "start", connectionMode, auditErr != nil)
-		s.auditHTTPEvent(r.Context(), PrincipalFromContext(r.Context()), providerName, "connection.oauth.start", auditAllowed, auditErr)
-	}()
-
-	var req startOAuthRequest
-	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
-		auditErr = errors.New("invalid JSON body")
-		writeError(w, http.StatusBadRequest, "invalid JSON body")
-		return
-	}
-	providerName = req.Integration
-
-	prov, ok := s.getProvider(w, req.Integration)
-	if !ok {
-		auditErr = errors.New("integration not found")
-		return
-	}
-	metricProviderName = req.Integration
-	connectionMode = metricutil.NormalizeConnectionMode(prov.ConnectionMode())
-
-	connection, ok := s.resolveRequestedConnection(w, req.Integration, req.Connection)
-	if !ok {
-		auditErr = errors.New("invalid connection")
-		return
-	}
-
-	handler, ok := s.requireOAuthHandler(w, req.Integration, connection)
-	if !ok {
-		auditErr = errors.New("oauth is not configured")
-		return
-	}
-
-	if s.stateCodec == nil {
-		auditErr = errors.New("oauth state encryption is not configured")
-		writeError(w, http.StatusInternalServerError, "oauth state encryption is not configured")
-		return
-	}
-
-	dbUserID, err := s.resolveUserID(w, r)
-	if err != nil {
-		auditErr = err
-		return
-	}
-
-	instance, ok := resolveRequestedInstance(w, req.Instance)
-	if !ok {
-		auditErr = errors.New("invalid instance")
-		return
-	}
-
-	connParams, ok := resolveConnectionParams(w, prov, req.ConnectionParams)
-	if !ok {
-		auditErr = errors.New("invalid connection parameters")
-		return
-	}
-
-	var (
-		authURL  string
-		verifier string
-	)
-
-	if len(connParams) > 0 {
-		rawAuthURL := handler.AuthorizationBaseURL()
-		resolvedAuthURL := paraminterp.Interpolate(rawAuthURL, connParams)
-		if resolvedAuthURL != rawAuthURL {
-			authURL, verifier = handler.StartOAuthWithOverride(resolvedAuthURL, "_", req.Scopes)
-		}
-	}
-	if authURL == "" {
-		authURL, verifier = handler.StartOAuth("_", req.Scopes)
-	}
-
-	authSource := ""
-	if p := PrincipalFromContext(r.Context()); p != nil {
-		authSource = p.AuthSource()
-	}
-	state, err := s.stateCodec.Encode(integrationOAuthState{
-		UserID:           dbUserID,
-		AuthSource:       authSource,
-		Integration:      req.Integration,
-		Connection:       connection,
-		Instance:         instance,
-		Verifier:         verifier,
-		ConnectionParams: connParams,
-		ExpiresAt:        s.now().Add(integrationOAuthStateTTL).Unix(),
-	})
-	if err != nil {
-		auditErr = errors.New("failed to encode oauth state")
-		writeError(w, http.StatusInternalServerError, "failed to encode oauth state")
-		return
-	}
-
-	authURL, err = setURLQueryParam(authURL, "state", state)
-	if err != nil {
-		auditErr = errors.New("failed to prepare oauth URL")
-		writeError(w, http.StatusInternalServerError, "failed to prepare oauth URL")
-		return
-	}
-
-	auditAllowed = true
-	auditErr = nil
-	writeJSON(w, http.StatusOK, map[string]string{"url": authURL, "state": state})
-}
-
-func (s *Server) integrationOAuthCallback(w http.ResponseWriter, r *http.Request) {
-	startedAt := time.Now()
-	auditAllowed := false
-	auditErr := errors.New("oauth callback failed")
-	auditUserID := ""
-	stateAuthSource := ""
-	providerName := ""
-	connectionMode := metricutil.UnknownAttrValue
-	defer func() {
-		metricutil.RecordConnectionAuthMetrics(r.Context(), startedAt, providerName, "oauth", "complete", connectionMode, auditErr != nil)
-		if auditUserID != "" {
-			s.auditHTTPEventWithUserID(r.Context(), auditUserID, stateAuthSource, providerName, "connection.oauth.complete", auditAllowed, auditErr)
-			return
-		}
-		s.auditHTTPEvent(r.Context(), nil, providerName, "connection.oauth.complete", auditAllowed, auditErr)
-	}()
-
-	writeCallbackError := func(status int, apiMessage, title, pageMessage string) {
-		if strings.Contains(strings.ToLower(r.Header.Get("Accept")), "text/html") {
-			writePendingConnectionPage(w, status, pendingConnectionPageView{
-				Title:     title,
-				Message:   pageMessage,
-				LinkURL:   "/integrations",
-				LinkLabel: "Open integrations",
-			}, "failed to render oauth error page")
-			return
-		}
-		writeError(w, status, apiMessage)
-	}
-
-	code := r.URL.Query().Get("code")
-	encodedState := r.URL.Query().Get("state")
-	if code == "" || encodedState == "" {
-		auditErr = errors.New("missing code or state parameter")
-		writeCallbackError(
-			http.StatusBadRequest,
-			"missing code or state parameter",
-			"Connection failed",
-			"The OAuth provider did not return the required callback parameters. Start the connection again from Integrations.",
-		)
-		return
-	}
-
-	if s.stateCodec == nil {
-		auditErr = errors.New("oauth state encryption is not configured")
-		writeCallbackError(
-			http.StatusInternalServerError,
-			"oauth state encryption is not configured",
-			"Connection failed",
-			"Gestalt could not validate this connection attempt. Contact your administrator and try again.",
-		)
-		return
-	}
-
-	state, err := s.stateCodec.Decode(encodedState, s.now())
-	if err != nil {
-		auditErr = errors.New("invalid or expired oauth state")
-		writeCallbackError(
-			http.StatusBadRequest,
-			"invalid or expired oauth state",
-			"Connection expired",
-			"This connection attempt is no longer valid. Start a new connection from Integrations.",
-		)
-		return
-	}
-	providerName = state.Integration
-	auditUserID = state.UserID
-	stateAuthSource = state.AuthSource
-	handler, ok := s.requireOAuthHandler(w, providerName, state.Connection)
-	if !ok {
-		auditErr = errors.New("oauth is not configured")
-		return
-	}
-
-	prov, _ := s.providers.Get(providerName)
-	if prov != nil {
-		connectionMode = metricutil.NormalizeConnectionMode(prov.ConnectionMode())
-	}
-
-	var exchangeOpts []oauth.ExchangeOption
-	connParams := state.ConnectionParams
-	if len(connParams) > 0 {
-		rawURL := handler.TokenURL()
-		resolved := paraminterp.Interpolate(rawURL, connParams)
-		if resolved != rawURL {
-			exchangeOpts = append(exchangeOpts, oauth.WithTokenURL(resolved))
-		}
-	}
-
-	var tokenResp *core.TokenResponse
-	tokenResp, err = handler.ExchangeCodeWithVerifier(r.Context(), code, state.Verifier, exchangeOpts...)
-	if err != nil {
-		auditErr = errors.New("token exchange failed")
-		slog.ErrorContext(r.Context(), "token exchange failed", "provider", providerName, "error", err)
-		writeCallbackError(
-			http.StatusBadGateway,
-			"token exchange failed",
-			providerName+" connection failed",
-			"The OAuth provider did not complete the connection. Start the connection again from Integrations.",
-		)
-		return
-	}
-
-	metadata, metaErr := buildConnectionMetadata(prov, connParams, tokenResp)
-	if metaErr != nil {
-		auditErr = errors.New("failed to extract connection metadata from token response")
-		slog.ErrorContext(r.Context(), "connection metadata extraction failed", "provider", providerName, "error", metaErr)
-		writeCallbackError(
-			http.StatusBadGateway,
-			"failed to extract connection metadata from token response",
-			providerName+" connection failed",
-			"Gestalt could not finish saving this connection. Start the connection again from Integrations.",
-		)
-		return
-	}
-
-	callbackInstance := state.Instance
-	if callbackInstance == "" {
-		callbackInstance = defaultTokenInstance
-	}
-
-	var tokenExpiresAt *time.Time
-	if tokenResp.ExpiresIn > 0 {
-		t := s.now().UTC().Truncate(time.Second).Add(time.Duration(tokenResp.ExpiresIn) * time.Second)
-		tokenExpiresAt = &t
-	}
-
-	tm := tokenMaterial{
-		UserID:         state.UserID,
-		AuthSource:     state.AuthSource,
-		Integration:    providerName,
-		Connection:     state.Connection,
-		Instance:       callbackInstance,
-		AccessToken:    tokenResp.AccessToken,
-		RefreshToken:   tokenResp.RefreshToken,
-		TokenExpiresAt: tokenExpiresAt,
-		MetadataJSON:   metadata,
-	}
-
-	result, err := s.runPostConnect(r.Context(), prov, tm)
-	if err != nil {
-		auditErr = errors.New("connection setup failed")
-		slog.ErrorContext(r.Context(), "post_connect failed", "provider", providerName, "error", err)
-		writeCallbackError(
-			http.StatusBadGateway,
-			"connection setup failed",
-			providerName+" connection failed",
-			"Gestalt could not finish saving this connection. Start the connection again from Integrations.",
-		)
-		return
-	}
-
-	if result.Status == "selection_required" {
-		state, err := s.decodePendingConnectionToken(result.PendingToken)
-		if err != nil {
-			auditErr = errors.New("failed to prepare pending connection")
-			writeError(w, http.StatusInternalServerError, "failed to prepare pending connection")
-			return
-		}
-		auditAllowed = true
-		auditErr = nil
-		s.writePendingConnectionSelectionPage(w, state, result.PendingToken)
-		return
-	}
-	auditAllowed = true
-	auditErr = nil
-	http.Redirect(w, r, "/integrations?connected="+url.QueryEscape(providerName), http.StatusSeeOther)
-}
-
-type connectManualRequest struct {
-	Integration      string            `json:"integration"`
-	Connection       string            `json:"connection"`
-	Instance         string            `json:"instance"`
-	Credential       string            `json:"credential"`
-	Credentials      map[string]string `json:"credentials"`
-	ConnectionParams map[string]string `json:"connection_params"`
-}
-
-func (s *Server) connectManual(w http.ResponseWriter, r *http.Request) {
-	startedAt := time.Now()
-	auditAllowed := false
-	auditErr := errors.New("manual connection failed")
-	providerName := ""
-	metricProviderName := metricutil.UnknownAttrValue
-	connectionMode := metricutil.UnknownAttrValue
-	defer func() {
-		metricutil.RecordConnectionAuthMetrics(r.Context(), startedAt, metricProviderName, "manual", "complete", connectionMode, auditErr != nil)
-		s.auditHTTPEvent(r.Context(), PrincipalFromContext(r.Context()), providerName, "connection.manual.connect", auditAllowed, auditErr)
-	}()
-
-	var req connectManualRequest
-	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
-		auditErr = errors.New("invalid JSON body")
-		writeError(w, http.StatusBadRequest, "invalid JSON body")
-		return
-	}
-	providerName = req.Integration
-
-	if req.Integration == "" {
-		auditErr = errors.New("integration is required")
-		writeError(w, http.StatusBadRequest, "integration is required")
-		return
-	}
-
-	prov, ok := s.getProvider(w, req.Integration)
-	if !ok {
-		auditErr = errors.New("integration not found")
-		return
-	}
-	metricProviderName = req.Integration
-	connectionMode = metricutil.NormalizeConnectionMode(prov.ConnectionMode())
-
-	mp, ok := prov.(core.ManualProvider)
-	if !ok || !mp.SupportsManualAuth() {
-		auditErr = errors.New("integration does not support manual auth")
-		writeError(w, http.StatusBadRequest, fmt.Sprintf("integration %q does not support manual auth; use OAuth connect instead", req.Integration))
-		return
-	}
-
-	dbUserID, err := s.resolveUserID(w, r)
-	if err != nil {
-		auditErr = err
-		return
-	}
-
-	manualInstance, ok := resolveRequestedInstance(w, req.Instance)
-	if !ok {
-		auditErr = errors.New("invalid instance")
-		return
-	}
-
-	manualConnection, ok := s.resolveRequestedConnection(w, req.Integration, req.Connection)
-	if !ok {
-		auditErr = errors.New("invalid connection")
-		return
-	}
-
-	auth := s.effectiveConnectionAuth(req.Integration, manualConnection)
-	effectiveCredential, credErr := buildEffectiveManualCredential(req, auth)
-	if credErr != nil {
-		auditErr = credErr
-		writeError(w, http.StatusBadRequest, credErr.Error())
-		return
-	}
-	if effectiveCredential == "" {
-		auditErr = errors.New("credential is required")
-		writeError(w, http.StatusBadRequest, "credential is required")
-		return
-	}
-
-	connParams, ok := resolveConnectionParams(w, prov, req.ConnectionParams)
-	if !ok {
-		auditErr = errors.New("invalid connection parameters")
-		return
-	}
-
-	manualMeta, metaErr := buildConnectionMetadata(prov, connParams, nil)
-	if metaErr != nil {
-		auditErr = errors.New(metaErr.Error())
-		writeError(w, http.StatusBadRequest, metaErr.Error())
-		return
-	}
-
-	authSource := ""
-	if p := PrincipalFromContext(r.Context()); p != nil {
-		authSource = p.AuthSource()
-	}
-	tm := tokenMaterial{
-		UserID:       dbUserID,
-		AuthSource:   authSource,
-		Integration:  req.Integration,
-		Connection:   manualConnection,
-		Instance:     manualInstance,
-		AccessToken:  effectiveCredential,
-		MetadataJSON: manualMeta,
-	}
-
-	result, err := s.runPostConnect(r.Context(), prov, tm)
-	if err != nil {
-		auditErr = errors.New("connection setup failed")
-		slog.ErrorContext(r.Context(), "post_connect failed", "provider", req.Integration, "error", err)
-		writeError(w, http.StatusBadGateway, "connection setup failed")
-		return
-	}
-
-	auditAllowed = true
-	auditErr = nil
-	writeJSON(w, http.StatusOK, result)
-}
-
-type createTokenRequest struct {
-	Name   string `json:"name"`
-	Scopes string `json:"scopes"`
-}
-
-type createTokenResponse struct {
-	ID        string     `json:"id"`
-	Name      string     `json:"name"`
-	Token     string     `json:"token"`
-	ExpiresAt *time.Time `json:"expires_at,omitempty"`
-}
-
-func (s *Server) createAPIToken(w http.ResponseWriter, r *http.Request) {
-	auditAllowed := false
-	auditErr := errors.New("api token creation failed")
-	defer func() {
-		s.auditHTTPEvent(r.Context(), PrincipalFromContext(r.Context()), "", "api_token.create", auditAllowed, auditErr)
-	}()
-
-	userID, err := s.resolveUserID(w, r)
-	if err != nil {
-		auditErr = err
-		return
-	}
-
-	var req createTokenRequest
-	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
-		auditErr = errors.New("invalid JSON body")
-		writeError(w, http.StatusBadRequest, "invalid JSON body")
-		return
-	}
-
-	if req.Name == "" {
-		auditErr = errors.New("name is required")
-		writeError(w, http.StatusBadRequest, "name is required")
-		return
-	}
-
-	if req.Scopes != "" {
-		for _, scope := range strings.Fields(req.Scopes) {
-			if _, err := s.providers.Get(scope); err != nil {
-				auditErr = fmt.Errorf("unknown scope %q", scope)
-				writeError(w, http.StatusBadRequest, fmt.Sprintf("unknown scope %q", scope))
-				return
-			}
-		}
-	}
-
-	apiToken, plaintext, err := s.issueAPIToken(r.Context(), userID, req.Name, req.Scopes, false)
-	if err != nil {
-		auditErr = errors.New("failed to generate token")
-		writeError(w, http.StatusInternalServerError, "failed to generate token")
-		return
-	}
-
-	auditAllowed = true
-	auditErr = nil
-	writeJSON(w, http.StatusCreated, createTokenResponse{
-		ID:        apiToken.ID,
-		Name:      apiToken.Name,
-		Token:     plaintext,
-		ExpiresAt: apiToken.ExpiresAt,
-	})
-}
-
-type apiTokenInfo struct {
-	ID        string     `json:"id"`
-	Name      string     `json:"name"`
-	Scopes    string     `json:"scopes"`
-	CreatedAt time.Time  `json:"created_at"`
-	ExpiresAt *time.Time `json:"expires_at,omitempty"`
-}
-
-func (s *Server) listAPITokens(w http.ResponseWriter, r *http.Request) {
-	userID, err := s.resolveUserID(w, r)
-	if err != nil {
-		return
-	}
-
-	tokens, err := s.datastore.ListAPITokens(r.Context(), userID)
-	if err != nil {
-		writeError(w, http.StatusInternalServerError, "failed to list tokens")
-		return
-	}
-
-	out := make([]apiTokenInfo, 0, len(tokens))
-	for _, t := range tokens {
-		out = append(out, apiTokenInfoFromCore(t))
-	}
-	writeJSON(w, http.StatusOK, out)
-}
-
-func (s *Server) integrationConnectionInfos(name string, integrationAuthTypes []string, defaultCredentialFields []credentialFieldInfo) []connectionDefInfo {
-	intg, ok := s.integrationDefs[name]
-	if !ok || intg.Plugin == nil {
-		return nil
-	}
-	return connectionInfosForPlugin(intg.Plugin, integrationAuthTypes, defaultCredentialFields)
-}
-
-func connectionInfosForPlugin(plugin *config.PluginDef, integrationAuthTypes []string, defaultCredentialFields []credentialFieldInfo) []connectionDefInfo {
-	if plugin == nil {
-		return nil
-	}
-	manifestProvider := plugin.ManifestPlugin()
-
-	var infos []connectionDefInfo
-	if info, ok := connectionInfoFromAuth(config.PluginConnectionAlias, config.EffectivePluginConnectionDef(plugin, manifestProvider).Auth, integrationAuthTypes, defaultCredentialFields); ok {
-		infos = append(infos, info)
-	}
-
-	names := make([]string, 0, len(plugin.Connections))
-	seen := make(map[string]struct{})
-	addName := func(raw string) {
-		resolved := config.ResolveConnectionAlias(raw)
-		if resolved == "" || resolved == config.PluginConnectionName {
-			return
-		}
-		if _, ok := seen[resolved]; ok {
-			return
-		}
-		seen[resolved] = struct{}{}
-		names = append(names, resolved)
-	}
-	if manifestProvider != nil {
-		for name := range manifestProvider.Connections {
-			addName(name)
-		}
-	}
-	for name := range plugin.Connections {
-		addName(name)
-	}
-	sort.Strings(names)
-
-	for _, name := range names {
-		conn, ok := config.EffectiveNamedConnectionDef(plugin, manifestProvider, name)
-		if !ok {
-			continue
-		}
-		if info, ok := connectionInfoFromAuth(name, conn.Auth, integrationAuthTypes, defaultCredentialFields); ok {
-			infos = append(infos, info)
-		}
-	}
-
-	return infos
-}
-
-func userFacingConnectionName(name string) string {
-	if name == config.PluginConnectionName {
-		return config.PluginConnectionAlias
-	}
-	return name
-}
-
-func integrationAuthTypesForProvider(prov core.Provider) []string {
-	var authTypes []string
-	if atl, ok := prov.(core.AuthTypeLister); ok {
-		authTypes = atl.AuthTypes()
-	} else if mp, ok := prov.(core.ManualProvider); ok && mp.SupportsManualAuth() {
-		authTypes = []string{"manual"}
-	} else {
-		authTypes = []string{"oauth"}
-	}
-	return userFacingAuthTypes(authTypes)
-}
-
-func credentialFieldInfosFromProvider(prov core.Provider) []credentialFieldInfo {
-	cfp, ok := prov.(core.CredentialFieldsProvider)
-	if !ok {
-		return nil
-	}
-	return credentialFieldInfos(cfp.CredentialFields(), func(field core.CredentialFieldDef) credentialFieldInfo {
-		return credentialFieldInfo{
-			Name:        field.Name,
-			Label:       field.Label,
-			Description: field.Description,
-			HelpURL:     field.HelpURL,
-		}
-	})
-}
-
-func credentialFieldInfos[T any](fields []T, mapField func(T) credentialFieldInfo) []credentialFieldInfo {
-	if len(fields) == 0 {
-		return nil
-	}
-	infos := make([]credentialFieldInfo, len(fields))
-	for i, field := range fields {
-		infos[i] = mapField(field)
-	}
-	return infos
-}
-
-func connectionInfoFromAuth(name string, auth config.ConnectionAuthDef, integrationAuthTypes []string, defaultCredentialFields []credentialFieldInfo) (connectionDefInfo, bool) {
-	authTypes := connectionAuthTypes(auth.Type, integrationAuthTypes)
-	if len(authTypes) == 0 {
-		return connectionDefInfo{}, false
-	}
-
-	info := connectionDefInfo{Name: name, AuthTypes: authTypes}
-	if fields := credentialFieldInfos(auth.Credentials, func(field config.CredentialFieldDef) credentialFieldInfo {
-		return credentialFieldInfo{
-			Name:        field.Name,
-			Label:       field.Label,
-			Description: field.Description,
-			HelpURL:     field.HelpURL,
-		}
-	}); len(fields) > 0 {
-		info.CredentialFields = fields
-	} else if authTypesContain(authTypes, "manual") && len(defaultCredentialFields) > 0 {
-		info.CredentialFields = append([]credentialFieldInfo(nil), defaultCredentialFields...)
-	}
-	return info, true
-}
-
-func (s *Server) effectiveConnectionAuth(integration, connection string) config.ConnectionAuthDef {
-	intg, ok := s.integrationDefs[integration]
-	if !ok || intg.Plugin == nil {
-		return config.ConnectionAuthDef{}
-	}
-	manifestProvider := intg.Plugin.ManifestPlugin()
-	if connection == config.PluginConnectionName {
-		return config.EffectivePluginConnectionDef(intg.Plugin, manifestProvider).Auth
-	}
-	conn, ok := config.EffectiveNamedConnectionDef(intg.Plugin, manifestProvider, connection)
-	if !ok {
-		return config.ConnectionAuthDef{}
-	}
-	return conn.Auth
-}
-
-func buildEffectiveManualCredential(req connectManualRequest, auth config.ConnectionAuthDef) (string, error) {
-	if len(req.Credentials) > 0 {
-		return marshalManualCredentials(req.Credentials)
-	}
-
-	structured := auth.AuthMapping != nil && (len(auth.AuthMapping.Headers) > 0 || auth.AuthMapping.Basic != nil)
-	if !structured {
-		return req.Credential, nil
-	}
-
-	switch {
-	case req.Credential != "" && len(auth.Credentials) == 1:
-		return marshalManualCredentials(map[string]string{auth.Credentials[0].Name: req.Credential})
-	case req.Credential != "":
-		return "", errors.New("manual connection requires named credentials")
-	}
-	return "", nil
-}
-
-func marshalManualCredentials(creds map[string]string) (string, error) {
-	if len(creds) == 0 {
-		return "", nil
-	}
-	for name, value := range creds {
-		if value == "" {
-			return "", fmt.Errorf("credential %q must not be empty", name)
-		}
-	}
-	data, err := json.Marshal(creds)
-	if err != nil {
-		return "", errors.New("invalid credentials map")
-	}
-	return string(data), nil
-}
-
-func connectionAuthTypes(authType string, integrationAuthTypes []string) []string {
-	if authType == "" {
-		if len(integrationAuthTypes) == 0 {
-			return nil
-		}
-		return append([]string(nil), integrationAuthTypes...)
-	}
-
-	authTypes := userFacingAuthTypes([]string{authType})
-	if len(authTypes) == 0 {
-		return nil
-	}
-	if authTypesContain(integrationAuthTypes, "manual") && !authTypesContain(authTypes, "manual") {
-		authTypes = append(authTypes, "manual")
-	}
-	return authTypes
-}
-
-func authTypesContain(authTypes []string, want string) bool {
-	for _, authType := range authTypes {
-		if authType == want {
-			return true
-		}
-	}
-	return false
-}
-
-func userFacingAuthTypes(authTypes []string) []string {
-	if len(authTypes) == 0 {
-		return nil
-	}
-	var out []string
-	for _, authType := range authTypes {
-		normalized, ok := userFacingAuthType(authType)
-		if !ok || authTypesContain(out, normalized) {
-			continue
-		}
-		out = append(out, normalized)
-	}
-	return out
-}
-
-func userFacingAuthType(authType string) (string, bool) {
-	switch authType {
-	case "", "oauth", "oauth2", "mcp_oauth":
-		return "oauth", true
-	case "manual", "bearer":
-		return "manual", true
-	default:
-		return "", false
-	}
-}
-
-func (s *Server) revokeAPIToken(w http.ResponseWriter, r *http.Request) {
-	auditAllowed := false
-	auditErr := errors.New("api token revoke failed")
-	defer func() {
-		s.auditHTTPEvent(r.Context(), PrincipalFromContext(r.Context()), "", "api_token.revoke", auditAllowed, auditErr)
-	}()
-
-	userID, err := s.resolveUserID(w, r)
-	if err != nil {
-		auditErr = err
-		return
-	}
-
-	id := chi.URLParam(r, "id")
-	if err := s.datastore.RevokeAPIToken(r.Context(), userID, id); err != nil {
-		if errors.Is(err, core.ErrNotFound) {
-			auditErr = errors.New("token not found")
-			writeError(w, http.StatusNotFound, "token not found")
-			return
-		}
-		auditErr = errors.New("failed to revoke token")
-		writeError(w, http.StatusInternalServerError, "failed to revoke token")
-		return
-	}
-	auditAllowed = true
-	auditErr = nil
-	writeJSON(w, http.StatusOK, map[string]string{"status": "revoked"})
-}
-
-func (s *Server) revokeAllAPITokens(w http.ResponseWriter, r *http.Request) {
-	auditAllowed := false
-	auditErr := errors.New("api token revoke all failed")
-	defer func() {
-		s.auditHTTPEvent(r.Context(), PrincipalFromContext(r.Context()), "", "api_token.revoke_all", auditAllowed, auditErr)
-	}()
-
-	userID, err := s.resolveUserID(w, r)
-	if err != nil {
-		auditErr = err
-		return
-	}
-
-	count, err := s.datastore.RevokeAllAPITokens(r.Context(), userID)
-	if err != nil {
-		auditErr = errors.New("failed to revoke tokens")
-		writeError(w, http.StatusInternalServerError, "failed to revoke tokens")
-		return
-	}
-	auditAllowed = true
-	auditErr = nil
-	writeJSON(w, http.StatusOK, map[string]any{"status": "revoked", "count": count})
-}
-
-func (s *Server) logout(w http.ResponseWriter, r *http.Request) {
-	auditAllowed := false
-	auditErr := errors.New("logout failed")
-	var auditPrincipal *principal.Principal
-	if !s.noAuth {
-		p, err := s.resolveRequestPrincipalWithUserID(r)
-		switch {
-		case err == nil:
-			auditPrincipal = p
-		case errors.Is(err, errInvalidAuthorizationHeader), errors.Is(err, principal.ErrInvalidToken):
-			slog.InfoContext(r.Context(), "logout: unable to resolve caller for audit", "error", err)
-		default:
-			slog.WarnContext(r.Context(), "logout: unable to resolve caller for audit", "error", err)
-		}
-	}
-	defer func() {
-		s.auditHTTPEvent(r.Context(), auditPrincipal, s.authProviderName(), "auth.logout", auditAllowed, auditErr)
-	}()
-
-	s.clearSessionCookie(w)
-	auditAllowed = true
-	auditErr = nil
-	writeJSON(w, http.StatusOK, map[string]string{"status": "ok"})
-}
-
-var (
-	safeParamValue         = regexp.MustCompile(`^[a-zA-Z0-9._-]+$`)
-	safeTokenResponseValue = regexp.MustCompile(`^[a-zA-Z0-9._:/-]+$`)
-)
-
-func validateConnectionParams(defs map[string]core.ConnectionParamDef, provided map[string]string) (map[string]string, error) {
-	if len(defs) == 0 {
-		return nil, nil
-	}
-	for key := range provided {
-		if _, ok := defs[key]; !ok {
-			return nil, fmt.Errorf("unknown connection parameter: %s", key)
-		}
-	}
-	result := make(map[string]string)
-	for name, def := range defs {
-		if def.From != "" {
-			continue
-		}
-		if v, ok := provided[name]; ok && v != "" {
-			if !safeParamValue.MatchString(v) {
-				return nil, fmt.Errorf("connection parameter %q contains invalid characters (allowed: letters, digits, hyphens, dots, underscores)", name)
-			}
-			result[name] = v
-		} else if def.Default != "" {
-			result[name] = def.Default
-		} else if def.Required {
-			return nil, fmt.Errorf("missing required connection parameter: %s", name)
-		}
-	}
-	if len(result) == 0 {
-		return nil, nil
-	}
-	return result, nil
-}
-
-func buildConnectionMetadata(prov core.Provider, userParams map[string]string, tokenResp *core.TokenResponse) (string, error) {
-	metadata := make(map[string]string)
-	for k, v := range userParams {
-		metadata[k] = v
-	}
-
-	if cpp, ok := prov.(core.ConnectionParamProvider); ok && tokenResp != nil && tokenResp.Extra != nil {
-		for name, def := range cpp.ConnectionParamDefs() {
-			if def.From == "token_response" {
-				field := def.Field
-				if field == "" {
-					field = name
-				}
-				val, ok := tokenResp.Extra[field]
-				if !ok {
-					if def.Required {
-						return "", fmt.Errorf("token response missing required field %q for connection param %q", field, name)
-					}
-					continue
-				}
-				s := fmt.Sprintf("%v", val)
-				if !safeTokenResponseValue.MatchString(s) {
-					return "", fmt.Errorf("token response field %q for connection param %q contains invalid characters", field, name)
-				}
-				metadata[name] = s
-			}
-		}
-	}
-
-	if len(metadata) == 0 {
-		return "", nil
-	}
-	b, err := json.Marshal(metadata)
-	if err != nil {
-		return "", fmt.Errorf("marshal connection metadata: %w", err)
-	}
-	return string(b), nil
-}
-
-type bearerTransport struct {
-	token string
-	base  http.RoundTripper
-}
-
-func (t *bearerTransport) RoundTrip(req *http.Request) (*http.Response, error) {
-	if req.Header.Get("Authorization") == "" {
-		req = req.Clone(req.Context())
-		req.Header.Set("Authorization", core.BearerScheme+t.token)
-	}
-	return t.base.RoundTrip(req)
-}
-
-type tokenMaterial struct {
-	UserID         string
-	AuthSource     string
-	Integration    string
-	Connection     string
-	Instance       string
-	AccessToken    string
-	RefreshToken   string
-	TokenExpiresAt *time.Time
-	MetadataJSON   string
-}
-
-type postConnectResult struct {
-	Status       string                   `json:"status"`
-	Integration  string                   `json:"integration,omitempty"`
-	SelectionURL string                   `json:"selection_url,omitempty"`
-	PendingToken string                   `json:"pending_token,omitempty"`
-	Candidates   []discoveryCandidateInfo `json:"candidates,omitempty"`
-}
-
-type discoveryCandidateInfo struct {
-	ID   string `json:"id"`
-	Name string `json:"name,omitempty"`
-}
-
-func (s *Server) storeTokenFromMaterial(ctx context.Context, tm tokenMaterial) (*core.IntegrationToken, error) {
-	now := s.now().UTC().Truncate(time.Second)
-	tok := &core.IntegrationToken{
-		ID:              uuid.NewString(),
-		UserID:          tm.UserID,
-		Integration:     tm.Integration,
-		Connection:      tm.Connection,
-		Instance:        tm.Instance,
-		AccessToken:     tm.AccessToken,
-		RefreshToken:    tm.RefreshToken,
-		ExpiresAt:       tm.TokenExpiresAt,
-		LastRefreshedAt: &now,
-		MetadataJSON:    tm.MetadataJSON,
-		CreatedAt:       now,
-		UpdatedAt:       now,
-	}
-	if err := s.datastore.StoreToken(ctx, tok); err != nil {
-		return nil, err
-	}
-	return tok, nil
-}
-
-func validateDiscoveryMetadata(metadata map[string]string) error {
-	for k, v := range metadata {
-		if !safeParamValue.MatchString(k) || !safeTokenResponseValue.MatchString(v) {
-			return fmt.Errorf("discovery returned invalid key or value for %q", k)
-		}
-	}
-	return nil
-}
-func mergeMetadataJSON(existing string, extra map[string]string) (string, error) {
-	m := make(map[string]string)
-	if existing != "" {
-		if err := json.Unmarshal([]byte(existing), &m); err != nil {
-			return "", fmt.Errorf("corrupt MetadataJSON: %w", err)
-		}
-	}
-	for k, v := range extra {
-		m[k] = v
-	}
-	b, err := json.Marshal(m)
-	if err != nil {
-		return "", fmt.Errorf("marshal merged metadata: %w", err)
-	}
-	return string(b), nil
-}
-
-func (s *Server) runPostConnect(ctx context.Context, prov core.Provider, tm tokenMaterial) (*postConnectResult, error) {
-	if dcp, ok := prov.(core.DiscoveryConfigProvider); ok {
-		if cfg := dcp.DiscoveryConfig(); cfg != nil {
-			client := &http.Client{
-				Timeout:   30 * time.Second,
-				Transport: &bearerTransport{token: tm.AccessToken, base: http.DefaultTransport},
-			}
-			candidates, err := discovery.Run(ctx, cfg, client)
-			if err != nil {
-				return nil, fmt.Errorf("discovery: %w", err)
-			}
-			if len(candidates) == 0 {
-				return nil, fmt.Errorf("no resources discovered")
-			}
-			if len(candidates) == 1 {
-				if err := validateDiscoveryMetadata(candidates[0].Metadata); err != nil {
-					return nil, err
-				}
-				merged, err := mergeMetadataJSON(tm.MetadataJSON, candidates[0].Metadata)
-				if err != nil {
-					return nil, err
-				}
-				tm.MetadataJSON = merged
-				if _, err := s.storeTokenFromMaterial(ctx, tm); err != nil {
-					return nil, err
-				}
-				return &postConnectResult{Status: "connected", Integration: tm.Integration}, nil
-			}
-
-			pendingToken, err := s.encodePendingConnectionToken(tm, candidates)
-			if err != nil {
-				return nil, fmt.Errorf("encode pending connection: %w", err)
-			}
-			return &postConnectResult{
-				Status:       "selection_required",
-				Integration:  tm.Integration,
-				SelectionURL: pendingConnectionPath,
-				PendingToken: pendingToken,
-				Candidates:   discoveryCandidateInfos(candidates),
-			}, nil
-		}
-	}
-
-	if _, err := s.storeTokenFromMaterial(ctx, tm); err != nil {
-		return nil, err
-	}
-	return &postConnectResult{Status: "connected", Integration: tm.Integration}, nil
-}
-
-func discoveryCandidateInfos(candidates []core.DiscoveryCandidate) []discoveryCandidateInfo {
-	if len(candidates) == 0 {
-		return nil
-	}
-	out := make([]discoveryCandidateInfo, len(candidates))
-	for i, candidate := range candidates {
-		out[i] = discoveryCandidateInfo{
-			ID:   candidate.ID,
-			Name: candidate.Name,
-		}
-	}
-	return out
 }
